Create simple web server
Adds support to run a simple webserver that currently only has a /health
endpoint. We use Express with helmet support to set important headers
along with CORS to only allow from specific origins, currently we allow
from all origins but this might change in the future.

We also create a specific error classes with some of the errors we might
come across such as duplicate emails or when an employee doesn't exists
or if validation fails at some point.

The webserver can be run using `npm run dev` and then we can query the
/health endpoint on `localhost:3000/health`