Add support of creating employee
This adds the route /api/employees/ that can be used to create an
employee. To do this we use the controller, service and repository
pattern.

The controller handles the incoming request, the service handles the
validation, checking that a user exists etc, and finally the repository
stores the data. This would ideally be stored in a database of some sort
but for now we just store it in memory.