Add support to fetch all employees
This adds a route /api/employees that will return all employees, we also
add support of pagination. This is good in the case that we have loads
of users and don't want to return all at once.

If you don't pass any query params we just return the first page and at
most 10 users, but you can also pass the query params page and limit
specify which determines which and how many users we return. You can
request between 1 and 100 users at a time.