Add Dockerfile
Adds a Dockerfile if we want to run the app using Docker, it's fairly
basic. We add a build step to build the app and then a production that
only installs the required packages.

A user nodejs is created and made the owner of the /app directory to
prevent anyone else running the app. We finally add a healthcheck that
checks the /health endpoint every 30s to see that the app is alive.